pyproject.toml
[tool.poetry]
name = "smartcalc"
version = "0.1.0"
description = ""
authors = ["Maxim Tsaregradtsev <3m3rcy3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.30.0"
fastapi = "^0.115.2"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
pydantic-settings = "^2.6.0"
uvicorn = "^0.32.0"
black = "^24.10.0"
fastapi-users-db-sqlalchemy = "^6.0.1"
redis = "^5.2.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
